Extract preview and recording controls
